refactor(web): extract dependencies-section from task detail
